requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coinmetrics-api-client==2023.5.17.19
idna==3.4
orjson==3.8.12
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typer==0.7.0
urllib3==2.0.2
websocket-client==1.5.2
